+++ NEWS.md
@@ -15,6 +15,8 @@
   (@jimhester, #1188)
 
 * `install_github()` now downloads tarballs rather than zipballs. (@kevinushey)
+* New `install_bioc()` function and bioc remote to install Bioconductor
+  packages from their SVN repository.
 
 * Suggested packages, including those specified as `Remotes:` are now installed
   after package installation. This allows you to use circular `Remotes:`